Update: 'songs' component folder renamed to 'tracks'
