+++ src/app/pages/album/Album.tsx
@@ -11,7 +11,7 @@ import NotFound from '../NotFound'
 
 import AlbumWrapper from './components/AlbumWrapper'
 import ActionBar from '../../../components/ActionBar'
-import TrackList from '../../../components/songs/TrackList'
+import TrackList from '../../../components/tracks/TrackList'
 import Footer from '../../../components/Footer'
 
 const Album = () => {
