+++ src/app/pages/artist/Artist.tsx
@@ -18,7 +18,7 @@ import ArtistWrapper from './components/ArtistWrapper'
 import { TabMenu, TabItems, TabTrigger, TabContent } from '../../../components/TabMenu'
 import BackgroundColour from '../../../components/BackgroundColour'
 import ActionBar from '../../../components/ActionBar'
-import TrackList from '../../../components/songs/TrackList'
+import TrackList from '../../../components/tracks/TrackList'
 import { Carousel, CarouselContainer, CarouselSlide } from '../../../components/carousel/Carousel'
 import ContentCard from '../../../components/content/ContentCard'
 import ContentSection from '../../../components/content/ContentSection'
