+++ src/app/pages/profile/Profile.tsx
@@ -13,7 +13,7 @@ import NotFound from '../NotFound'
 import ProfileWrapper from './components/ProfileWrapper'
 import { TabContent, TabItems, TabMenu, TabTrigger } from '../../../components/TabMenu'
 import BackgroundColour from '../../../components/BackgroundColour'
-import TrackList from '../../../components/songs/TrackList'
+import TrackList from '../../../components/tracks/TrackList'
 import { Carousel, CarouselContainer, CarouselSlide } from '../../../components/carousel/Carousel'
 import ContentCard from '../../../components/content/ContentCard'
 import Footer from '../../../components/Footer'
